+++ expected_response/lib/src/annotations/request.dart
@@ -1,0 +1,6 @@
+/// Contains the @EntryFunction annotation
+/// 
+/// This annotation represents the function that will be evoked to make a request to the server
+class Request {
+  const Request();
+}