+++ expected_response/lib/src/annotations/response.dart
@@ -2,7 +2,7 @@
 /// 
 /// This define the behaviour for each response code in the return of an http request
 class Response {
-  const Response({this.code});
+  const Response([this.code]);
 
   /// The return code from an http response
   final int code;