+++ expected_response/lib/src/exceptions/unrecognized_http_exception.dart
@@ -1,0 +1,8 @@
+import 'package:expected_response/src/server_response.dart';
+
+/// Is the default exception for an unknown status code
+class UnrecognizedHttpResponse implements Exception {
+  UnrecognizedHttpResponse(this.response);
+
+  final ServerResponse response;
+}