+++ expected_response_generator/lib/expected_response_generator.dart
@@ -1,0 +1,1 @@
+export 'src/expected_response_generator.dart' show ExpectResponseGenerator;