+++ http_response/lib/src/annotations/catch_error.dart
@@ -1,6 +1,0 @@
-/// Contains the @CatchError annotation
-/// 
-/// This annotation represent the behaviour of the response when an exception is throw
-class CatchError {
-  const CatchError();
-}