+++ http_response/lib/src/annotations/exception_response.dart
@@ -1,3 +1,0 @@
-class ExceptionResponse {
-  const ExceptionResponse();
-}