+++ http_response/lib/src/annotations/http_response.dart
@@ -1,6 +1,0 @@
-/// Contains the @HttpResponse annotation
-/// 
-/// This annotations is the entrypoint of everything, identifies the class where the ecossystem will run
-class HttpResponse {
-  const HttpResponse();
-}