+++ http_response/lib/src/server_response.dart
@@ -1,3 +1,0 @@
-abstract class ServerResponse {
-  int get statusCode;
-}