+++ expected_response/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 - Create package
 - Create the annotations
-    - HttpResponse
+    - ExpectedResponse
         - Responsable for indicate the class where the ecossystem will run, is the entry point of everything;
     - Response
         - Identify the behaviour expected for each response code returned from the http response
+    - Request
+        - Identify the function that must make the request
+    - EntryFunction
+        - Is the function that will evoke the generated function
 - Create ServerResponse, the abstract class that receives the http response
+- Created the UnrecognizedHttpResponse, the default behaviour for unknown responses
+- Created the HttpBase, the super class for every class annotated with @ExpectedResponse
 - Documented the annotations
 - Documented the helper classes 
 - Create README